Fixed issue with ezformtoken extension, file type filter and location to flash and silverlight files
I have created a custom template operator to pull in the form token so
that it can be passed with the file uploads to handle ezformtoken being
installed.

I have also fixed up some the location of the silverlight and flash
files.

Fixed the file type filter output so that it works properly now